Pushing even more of the generic parts of the callbacks into the firebase fascade
